+++ ft_printf/trees/flags_disp_tree.c
@@ -6,7 +6,7 @@
 /*   By: mlabouri <mlabouri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/12 16:14:08 by mlabouri          #+#    #+#             */
-/*   Updated: 2020/02/10 09:41:31 by mlabouri         ###   ########.fr       */
+/*   Updated: 2020/02/10 09:41:42 by mlabouri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ void	disp_minus(int bin[], unsigned long long int arg, char conv)
 	if (ft_cinset(conv, INTS))
 		length = bin['-'] - length_int(conv, arg);
 	else
-		length = bin['-'] - func[conv](conv, arg);
+		length = bin['-'] - func[(int)conv](conv, arg);
 	ft_disp_convs(arg, conv);
 	while (length > 0)
 	{
@@ -46,7 +46,7 @@ void	disp_zero(int bin[], unsigned long long int arg, char conv)
 	if (ft_cinset(conv, INTS))
 		length = bin['0'] - length_int(conv, arg);
 	else
-		length = bin['0'] - func[conv](conv, arg);
+		length = bin['0'] - func[(int)conv](conv, arg);
 	if (ft_cinset(conv, "di") && (int)arg < 0)
 		arg = (unsigned long long)int_abs((int)arg);
 	while (length > 0)
@@ -91,7 +91,7 @@ void	disp_pad(int bin[], unsigned long long int arg, char conv)
 	if (ft_cinset(conv, INTS))
 		length = bin[' '] - length_int(conv, arg);
 	else
-		length = bin[' '] - func[conv](conv, arg);
+		length = bin[' '] - func[(int)conv](conv, arg);
 	while (length > 0)
 	{
 		ft_putchar_fd(' ', 1);
